+++ etlrules/backends/common/newcolumns.py
@@ -80,3 +80,55 @@ class AddNewColumnRule(UnaryOpBaseRule):
     def _validate_columns(self, df_columns):
         if self.strict and self.output_column in df_columns:
             raise ColumnAlreadyExistsError(f"Column {self.output_column} already exists in the input dataframe.")
+
+
+class AddRowNumbersRule(UnaryOpBaseRule):
+    """ Adds a new column with row numbers.
+
+    Example::
+
+        Given df:
+        | A   | B  |
+        | 1   | 2  |
+        | 2   | 3  |
+        | 3   | 4  |
+
+    > AddRowNumbersRule("Row_Number").apply(df)
+
+    Result::
+
+        | A   | B  | Row_Number |
+        | 1   | 2  | 0          |
+        | 2   | 3  | 1          |
+        | 3   | 4  | 2          |
+
+    Args:
+        output_column: The name of the new column to be added.
+        start: The value to start the numbers from. Defaults to 0.
+        step: The increment to be used between row numbers. Defaults to 1.
+
+        named_input: Which dataframe to use as the input. Optional.
+            When not set, the input is taken from the main output.
+            Set it to a string value, the name of an output dataframe of a previous rule.
+        named_output: Give the output of this rule a name so it can be used by another rule as a named input. Optional.
+            When not set, the result of this rule will be available as the main output.
+            When set to a name (string), the result will be available as that named output.
+        name: Give the rule a name. Optional.
+            Named rules are more descriptive as to what they're trying to do/the intent.
+        description: Describe in detail what the rules does, how it does it. Optional.
+            Together with the name, the description acts as the documentation of the rule.
+        strict: When set to True, the rule does a stricter valiation. Default: True
+
+    Raises:
+        ColumnAlreadyExistsError: raised in strict mode only if a column with the same name already exists in the dataframe.
+    """
+
+    def __init__(self, output_column: str, start: int=0, step: int=1, named_input: Optional[str]=None, named_output: Optional[str]=None, name: Optional[str]=None, description: Optional[str]=None, strict: bool=True):
+        super().__init__(named_input=named_input, named_output=named_output, name=name, description=description, strict=strict)
+        self.output_column = output_column
+        self.start = start
+        self.step = step
+
+    def _validate_columns(self, df_columns):
+        if self.strict and self.output_column in df_columns:
+            raise ColumnAlreadyExistsError(f"Column {self.output_column} already exists in the input dataframe.")