+++ etlrules/backends/pandas/newcolumns.py
@@ -1,4 +1,9 @@
-from etlrules.backends.common.newcolumns import AddNewColumnRule as AddNewColumnRuleBase
+import numpy as np
+
+from etlrules.backends.common.newcolumns import (
+    AddNewColumnRule as AddNewColumnRuleBase,
+    AddRowNumbersRule as AddRowNumbersRuleBase,
+)
 from etlrules.backends.pandas.expressions import Expression
 from etlrules.backends.pandas.types import MAP_TYPES
 
@@ -16,5 +21,16 @@ class AddNewColumnRule(AddNewColumnRuleBase):
                 result = result.astype(MAP_TYPES[self.column_type])
             except ValueError as exc:
                 raise TypeError(str(exc))
+        df = df.assign(**{self.output_column: result})
+        self._set_output_df(data, df)
+
+
+class AddRowNumbersRule(AddRowNumbersRuleBase):
+
+    def apply(self, data):
+        df = self._get_input_df(data)
+        self._validate_columns(df.columns)
+        stop = self.start + df.shape[0] * self.step
+        result = np.arange(start=self.start, stop=stop, step=self.step)
         df = df.assign(**{self.output_column: result})
         self._set_output_df(data, df)