+++ etlrules/backends/polars/newcolumns.py
@@ -1,6 +1,9 @@
 import polars as pl
 
-from etlrules.backends.common.newcolumns import AddNewColumnRule as AddNewColumnRuleBase
+from etlrules.backends.common.newcolumns import (
+    AddNewColumnRule as AddNewColumnRuleBase,
+    AddRowNumbersRule as AddRowNumbersRuleBase,
+)
 from etlrules.backends.polars.expressions import Expression
 from etlrules.backends.polars.types import MAP_TYPES
 
@@ -23,4 +26,14 @@ class AddNewColumnRule(AddNewColumnRuleBase):
             except pl.exceptions.ComputeError as exc:
                 raise TypeError(exc)
         df = df.with_columns(**{self.output_column: result})
+        self._set_output_df(data, df)
+
+
+class AddRowNumbersRule(AddRowNumbersRuleBase):
+
+    def apply(self, data):
+        df = self._get_input_df(data)
+        self._validate_columns(df.columns)
+        stop = self.start + len(df) * self.step
+        df = df.with_columns(**{self.output_column: pl.arange(start=self.start, end=stop, step=self.step)})
         self._set_output_df(data, df)